bnc#1031106 - SLE-12-SP2-SAP-ppc64le-GM-DVD1-rebuild.iso | -bash: xdg-desktop-icon: command not found error before starting yast.ssh to install

+++ src/misc/sap-profile.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [ ! -e $HOME/.sap-icons ]; then 
+if [ -e /usr/bin/xdg-desktop-icon -a ! -e $HOME/.sap-icons ]; then 
 	xdg-desktop-icon install /usr/share/applications/YaST2/sap-windows_cheat_sheet.desktop
 	xdg-desktop-icon install /usr/share/applications/YaST2/sap-installation-wizard.desktop
 	xdg-desktop-icon install --novendor /usr/share/applications/YaST2/customer_center.desktop
